[FIX] `extractToken` 반환값을 변수에 저장하여 검증 오류 수정
`Bearer`가 포함된 채로 `validateToken`에 전달되는 문제 해결

Related to: TICO-162, TICO-206, TICO-246, TICO-241
